requirements.txt
robotframework
setuptools
robotframework-browser
robotframework-requests
pyyaml
robotframework-faker
robotframework-debuglibrary
beautifulsoup4
pyzbar
pillow